requirements.txt
firecrawl-py==2.7.0
openai==1.35.0
httpx==0.25.0
beautifulsoup4==4.12.3
pandas==2.2.2
requests==2.32.3
python-dotenv==1.0.1
tavily-python==0.3.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0 